backend/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0        # async PostgreSQL driver (used by SQLAlchemy async engine)
psycopg2-binary>=2.9.9 # sync driver (used by Alembic migrations if needed)

# Environment variables
python-dotenv>=1.0.0

# HTTP client (for Ollama)
httpx>=0.27.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.17.0

# Pydantic (bundled with FastAPI but pinned for safety)
pydantic>=2.7.0
